+++ tests/test_resource.py
